+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * added configuration file
 * added script integration
+* added A/B testing
+* added config file to Docker
 
 ## 0.1